pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hydra-inference"
version = "0.1.0"
description = "Mixed-vendor GPU inference cluster manager using llama.cpp RPC"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0",
    "click>=8.0",
    "httpx>=0.25",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21"]

[project.scripts]
hydra = "hydra.cli:cli"

[tool.setuptools.packages.find]
include = ["hydra*"]
